extracted enums FilterType and ComperatorFilterOption, moved CustomFilterTest to AMW_business
