+++ AMW_business/src/main/java/ch/puzzle/itc/mobiliar/business/deploy/entity/ComperatorFilterOption.java
@@ -1,0 +1,18 @@
+package ch.puzzle.itc.mobiliar.business.deploy.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public enum ComperatorFilterOption {
+    smaller("<", " < ", null, null),
+    smallerequals("<=", " <= ", null, null),
+    equals("is", " = ", " like ", " is "),
+    greaterequals(">=", " >= ", null, null),
+    greater(">", " > ", null, null), notequal("is not", " != ", null, " is not ");
+    private String displayName;
+    private String sqlNumComperator;
+    private String sqlStringComperator;
+    private String sqlBoolComperator;
+}
