+++ AMW_business/src/main/java/ch/puzzle/itc/mobiliar/business/deploy/entity/FilterType.java
@@ -1,0 +1,5 @@
+package ch.puzzle.itc.mobiliar.business.deploy.entity;
+
+public enum FilterType {
+    booleanType, StringType, IntegerType, DateType, SpecialFilterType, LabeledDateType, ENUM_TYPE
+}
